pyproject.toml
[tool.poetry]
name = "siwe"
version = "1.0.0"
description = "A Python implementation of Sign-In with Ethereum (EIP-4361)."
license = "MIT OR Apache-2.0"
authors = ["Spruce Systems, Inc. <hello@spruceid.com>", "Payton Garland <payton.r.g@gmail.com>"]
readme = "README.md"
homepage = "https://login.xyz"
repository = "https://github.com/spruceid/siwe-py"
keywords = ["SIWE", "EIP-4361", "Sign-In with Ethereum", "Spruce ID"]

[tool.poetry.dependencies]
python = ">3.7.0,<3.7.1 || >3.7.1,<4.0"
abnf = "1.1.1"
web3 = "^5.26.0"
python-dateutil = "2.8.2"
eth-account = "^0.5.6"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^21.12b0"
pyhumps = "^3.5.0"
eth-account = "^0.5.6"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"EIP" = "https://github.com/ethereum/EIPs/blob/master/EIPS/eip-4361.md"
"Discord" = "https://discord.gg/Sf9tSFzrnt"
